Passions screen added, minor folder cleanup.
