+++ app/src/androidTest/java/ch/epfl/sdp/blindly/profile_preferences/TestProfile5.kt
@@ -1,15 +1,16 @@
-package ch.epfl.sdp.blindly
+package ch.epfl.sdp.blindly.profile_preferences
 
 import androidx.test.espresso.Espresso
 import androidx.test.espresso.action.ViewActions.click
 import androidx.test.espresso.intent.Intents
 import androidx.test.espresso.intent.Intents.intended
-import androidx.test.espresso.intent.matcher.IntentMatchers
 import androidx.test.espresso.intent.matcher.IntentMatchers.hasComponent
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import org.hamcrest.Matchers
+import ch.epfl.sdp.blindly.Profile5
+import ch.epfl.sdp.blindly.Profile6
+import ch.epfl.sdp.blindly.R
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
