+++ app/src/androidTest/java/ch/epfl/sdp/blindly/profile_preferences/TestProfile6.kt
@@ -1,34 +1,35 @@
-package ch.epfl.sdp.blindly
+package ch.epfl.sdp.blindly.profile_preferences
 
 import androidx.test.espresso.Espresso
 import androidx.test.espresso.action.ViewActions.click
 import androidx.test.espresso.intent.Intents
 import androidx.test.espresso.intent.Intents.intended
-import androidx.test.espresso.intent.matcher.IntentMatchers
 import androidx.test.espresso.intent.matcher.IntentMatchers.hasComponent
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import org.hamcrest.Matchers
+import ch.epfl.sdp.blindly.Profile6
+import ch.epfl.sdp.blindly.Profile7
+import ch.epfl.sdp.blindly.R
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
 
 
 @RunWith(AndroidJUnit4::class)
-class TestProfile4 {
+class TestProfile6 {
 
 
     @get:Rule
-    val activityRule = ActivityScenarioRule(Profile4::class.java)
+    val activityRule = ActivityScenarioRule(Profile6::class.java)
 
 
     @Test
-    fun testProfile4FiresProfile5() {
+    fun testProfile6FiresProfile7() {
         Intents.init()
-        val buttonStart = Espresso.onView(withId(R.id.button_p4))
+        val buttonStart = Espresso.onView(withId(R.id.button_p6))
         buttonStart.perform(click())
-        intended(hasComponent(Profile5::class.java.name))
+        intended(hasComponent(Profile7::class.java.name))
         Intents.release()
     }
