+++ app/src/main/java/ch/epfl/sdp/blindly/Profile7.kt
@@ -1,0 +1,11 @@
+package ch.epfl.sdp.blindly
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+
+class Profile7 : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.set_profile_7)
+    }
+}
