+++ app/src/main/res/layout/set_profile_2.xml
@@ -57,7 +57,7 @@
         style="?android:attr/progressBarStyleHorizontal"
         android:layout_width="413dp"
         android:layout_height="62dp"
-        android:progress="20"
+        android:progress="17"
         app:layout_constraintBottom_toTopOf="@+id/text_p2"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.0"
