+++ app/src/main/res/layout/set_profile_3.xml
@@ -23,7 +23,7 @@
         style="?android:attr/progressBarStyleHorizontal"
         android:layout_width="413dp"
         android:layout_height="62dp"
-        android:progress="40"
+        android:progress="34"
         app:layout_constraintBottom_toTopOf="@+id/text_p3"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.0"
