+++ app/src/main/res/layout/set_profile_4.xml
@@ -37,7 +37,7 @@
         style="?android:attr/progressBarStyleHorizontal"
         android:layout_width="413dp"
         android:layout_height="62dp"
-        android:progress="60"
+        android:progress="51"
         app:layout_constraintBottom_toTopOf="@+id/text_p4"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.0"
