+++ app/src/main/res/layout/set_profile_4_2.xml
@@ -11,10 +11,9 @@
         style="?android:attr/progressBarStyleHorizontal"
         android:layout_width="413dp"
         android:layout_height="62dp"
-        android:progress="60"
-        app:layout_constraintBottom_toTopOf="@+id/text_p4"
+        android:progress="51"
+        app:layout_constraintBottom_toTopOf="@+id/text_p4_2"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintVertical_bias="0.0" />
